fixups

+++ src/parameters.cu
@@ -100,7 +100,7 @@ void Parameters::get_nml_vars(char* filename)
   tnml = nml;  
   if (nml.contains("Time")) tnml = toml::find (nml, "Time");
   dt      = toml::find_or <float> (tnml, "dt",       0.05 );
-  dt_max  = toml::find_or <float> (tnml, "dt_max",   dt );
+  dt_max  = toml::find_or <float> (tnml, "dt_max",   static_cast<float>(dt) );
   dt_min  = toml::find_or <float> (tnml, "dt_min",   1e-7 );
   fixed_dt = toml::find_or <bool> (tnml, "fixed_dt", false );
